First full permission implementation
A small explanation:

Each permission has a "precision" to it. The permission with the
highest precision for any given case "wins".
Wildcards are of course less precise than direct permissions (aptly
called 'Precise'), Wildcards nearer to the root are less precise than
those closer to the searched element.

Each permission can be either positive (that is 'grant') or negative
(that is 'deny'). Positive and Negative permissions are not ordered,
having two conflicting permissions with the same precision is undefined
behaviour and thus must be checked for in config parsing or any
permission-changing interface.